Move preprocessing out of the model into seperate CLI command
