+++ tests/test_raster.py
@@ -1,7 +1,9 @@
 import msprime
 import pytest
+import tszip
 
 from tsbrowse import pages
+from tsbrowse import preprocess
 from tsbrowse import raster
 
 
@@ -17,11 +19,12 @@ def ts():
 
 class TestRaster:
     @pytest.mark.parametrize("page", pages.PAGES_MAP.values())
-    def test_mutation_scatter(self, page, ts, tmp_path):
-        print(ts)
+    def test_mutation_scatter(self, page, ts, tmp_path, tmpdir):
+        tszip.compress(ts, tmpdir / "test.tszip")
+        preprocess.preprocess(tmpdir / "test.tszip", tmpdir / "test.tsbrowse")
         raster.raster_component(
             page.page,
-            ts,
+            tmpdir / "test.tsbrowse",
             tmp_path / "image.png",
         )
         assert (tmp_path / "image.png").exists()
